started possible workaround lang tag handling
Looks like it's not possible to get dynamic lang tag from SQL column, thus,
the only way is to create partition per lang tag. This on the other needs for
post processing the resulting dataframe as appending the lang tag isn't
done in the generated SQL query but only in the Ontop framework afterwards
(which isn't serializable).
The post processing is still missing - Ontop substitution pipeline is
very complex ...